Revert "Hs 22 add handler for two foto open and close shift (#24)"
This reverts commit f662ceecf46f69f8c23dae8eede6d6e85abaa8e4.
