+++ .gitignore
@@ -2,4 +2,3 @@ poetry.lock
 .venv
 .idea
 *.lock
-/app/data/hookah_space.sqlite3
