+++ alembic.ini
@@ -3,7 +3,7 @@
 [alembic]
 # path to migration scripts.
 # Use forward slashes (/) also on windows to provide an os agnostic path
-script_location = migration
+script_location = db/migration
 
 # template used to generate migration file names; The default value is %%(rev)s_%%(slug)s
 # Uncomment the line below if you want the files to be prepended with date and time
