+++ fsm/shift_fsm.py
@@ -6,20 +6,15 @@ class ShiftStates(StatesGroup):
     choose_point = State()
     enter_cash = State()
     confirm_cash = State()
-    upload_light_tobacco_photo = State()
-    confirm_light_tobacco_photo = State()
-    upload_dark_tobacco_photo = State()
-    confirm_dark_tobacco_photo = State()
-    resume = State()
+    upload_tobacco_photo = State()
+    confirm_tobacco_photo = State()
     working = State()
 
     # Закрытие смены
     enter_cash_report = State()
     enter_terminal_report = State()
-    upload_remaining_light_tobacco = State()
-    confirm_remaining_light_tobacco_photo = State()
-    upload_remaining_dark_tobacco_photo = State()
-    confirm_remaining_dark_tobacco_photo = State()
+    upload_remaining_tobacco = State()
+    confirm_remaining_tobacco_photo = State()
     extra_information = State()
     confirm_close_shift = State()
