--- migration/versions/6f03c4d03223_add_dark_tobacco_photo_id_column_to_.py
@@ -1,34 +1,0 @@
-"""add dark tobacco photo id column to Shift table
-
-Revision ID: 6f03c4d03223
-Revises: ff944a10dee1
-Create Date: 2025-01-03 17:39:42.074543
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6f03c4d03223'
-down_revision: Union[str, None] = 'ff944a10dee1'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('shifts', sa.Column('start_shift_light_tobacco_photo_id', sa.Text(), nullable=True))
-    op.add_column('shifts', sa.Column('start_shift_dark_tobacco_photo_id', sa.Text(), nullable=True))
-    op.drop_column('shifts', 'start_shift_tobacco_photo_id')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('shifts', sa.Column('start_shift_tobacco_photo_id', sa.TEXT(), nullable=False))
-    op.drop_column('shifts', 'start_shift_dark_tobacco_photo_id')
-    op.drop_column('shifts', 'start_shift_light_tobacco_photo_id')
-    # ### end Alembic commands ###
