+++ routers/commands/start_commands.py
@@ -1,9 +1,3 @@
-from aiogram import Router, types
-from aiogram.filters import Command
+from aiogram import Router
 
 router = Router()
-
-
-@router.message(Command("start"))
-async def start(message: types.Message):
-    await message.answer("Привет! Я POS бот для <b>Shisha Space</b> .\nДля начала работы введите /start_shift")
